% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iter_library.R
\name{iter_library}
\alias{iter_library}
\title{Pass in a character Vector of Package Names to install and load into the Environment.}
\usage{
iter_library(package_names, continue_on_failed_install = FALSE)
}
\arguments{
\item{package_names}{character vector.  Names of the packages to install and load into the Environment.}

\item{continue_on_failed_install}{Boolean.  Stops R if a packages fails to install and parameter is set to FALSE.}
}
\value{
Boolean.  FALSE if no packages are passed into the library function else TRUE.
}
\description{
Pass in a character Vector of Package Names to install and load into the Environment.
}
\examples{

\dontrun{
 # Character Vector of packages to load and check to install
 packages_to_load <- c("consolechoice", "dplyr", "ggplot2")

 iter_library(packages_to_load)

 # R can continue working if a new package fails to load
 # by setting continue_on_failed_install to TRUE
 packages_that_might_not_load <- c("consolechoice", "packageThatDoesNotExist")

 iter_library(packages_that_might_not_load, continue_on_failed_install = TRUE)

}
}
