% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msg_wrap.R
\name{msg_wrap}
\alias{msg_wrap}
\title{Wraps a message before and after a function}
\usage{
msg_wrap(
  func,
  ...,
  before_func_msg = "",
  after_func_msg = "",
  print_func = print,
  use_msg = "both"
)
}
\arguments{
\item{func}{function}

\item{...}{Additional arguments to be passed to the param func.}

\item{before_func_msg}{character.}

\item{after_func_msg}{character.}

\item{print_func}{function the default is print.  Can use related function like message.}

\item{use_msg}{character.  Selects which messages to print in the function.  Use \code{before}, \code{after} or \code{both}.}
}
\value{
Unknown.  The return type from the param func.
}
\description{
Wraps a message before and after a function
}
\examples{

library(wrappr)

numbers <- c(1,2,3,4,5)

# wrapping a message before and after the sum function
answer <- msg_wrap(
                   sum,
                   numbers,
                   before_func_msg = "Currently summing the numbers",
                   after_func_msg = "Summing the numbers complete"
                   )

numbers_with_na <- c(1,2,3,NA,5)

# Summing with message only at start and NA removed
answer_na_removed <- msg_wrap(
                              sum,
                              numbers,
                              na.rm = TRUE,
                              before_func_msg = "Sum with na.rm set to TRUE",
                              use_msg = "before"
                              )
}
