% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cache_or_create.R
\name{get_cache_or_create}
\alias{get_cache_or_create}
\title{checks if variable exists in enviroment and returns back or creates a new variable}
\usage{
get_cache_or_create(
  var,
  func,
  ...,
  err_msg = "An error has occured in the function get_cache_or_create"
)
}
\arguments{
\item{var}{character.  The name of the variable to check in the global enviroment.}

\item{func}{function.  A function that returns a value.}

\item{...}{Additional arguments to be passed to the param func.}

\item{err_msg}{character.  Message sent to stop function if an error occurs.}
}
\value{
Unknown.  The return type from the param func or the existing variable in global enviroment.
}
\description{
checks if variable exists in enviroment and returns back or creates a new variable
}
