% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy_eval.R
\name{lazy_eval}
\alias{lazy_eval}
\title{save and Delay a function call with the option to change the function and args when called}
\usage{
lazy_eval(..., .f)
}
\arguments{
\item{...}{Additional arguments to be passed to the param .f.  Also in closure function returned.}

\item{.f}{function.  A function that will be called when needed.  Also in closure function returned.}
}
\value{
closure function with same function params plus an extra boolean param overwright_args set to FALSE.
}
\description{
save and Delay a function call with the option to change the function and args when called
}
\examples{

# perform simple calculations using the same data
numbers <- c(1,2,3,4,5)

func <- lazy_eval(numbers, .f = sum)

sum_result <- func()

max_result <- func(.f = max)

mean_result <- func(.f = mean)

range_result <- func(.f = function(...) { max(...) - min(...)})

add_more_num_result <- func(4,5,6, NA, na.rm = TRUE)

# overwrite_args set to TRUE will replace the original params for the function
new_num_result <- func(4,5,6, NA, na.rm = TRUE, overwrite_args = TRUE)

}
