% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy_eval.R
\name{lazy_eval}
\alias{lazy_eval}
\title{Delay a function call until you need to call it}
\usage{
lazy_eval(func, ...)
}
\arguments{
\item{func}{function.  A function that will be called when evoked later}

\item{...}{Additional arguments to be passed to the param func.}
}
\value{
function
}
\description{
Delay a function call until you need to call it
}
\examples{
\dontrun{
# Setup to load a csv file

load_file_two <- TRUE

csv_one <- lazy_eval(
                     read.csv,
                     file = "file/address/of/file_one.csv
                    )

csv_two <- lazy_eval(
                     read.csv,
                     file = "file/address/of/file_two.csv
                    )

data <- ifelse(load_file_two, csv_one(), csv_two())
}

}
